chore: refactor database classes
